requirements.txt
djangorestframework==3.13.1
Django==4.1
flake8==5.0.4
drf_yasg==1.21.4
django-filter==22.1


